requirements.txt
torch>=2.0
torchdiffeq>=0.2.3
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
scipy>=1.11
matplotlib>=3.7
plotly>=5.15
mlflow>=2.8
jupyter>=1.0
kaggle>=1.5
pyyaml>=6.0
